+++ src/main/java/io/github/astrapi69/crypto/model/KeyModel.java
@@ -1,0 +1,19 @@
+package io.github.astrapi69.crypto.model;
+
+import lombok.AccessLevel;
+import lombok.Data;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import lombok.experimental.SuperBuilder;
+
+@Data
+@RequiredArgsConstructor
+@SuperBuilder(toBuilder = true)
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
+public class KeyModel {
+    @NonNull
+    byte[] encoded;
+    @NonNull
+    String algorithm;
+}